Fix phpunit paths
