+++ tests/src/Mooc/Shared/Infrastructure/PhpUnit/MoocContextInfrastructureTestCase.php
@@ -4,7 +4,7 @@ declare(strict_types = 1);
 
 namespace CodelyTv\Tests\Mooc\Shared\Infrastructure\PhpUnit;
 
-use CodelyTv\Tests\Shared\Infrastructure\PhpUnit\InfrastructureTestCase;
+use CodelyTv\Tests\Shared\Infrastructure\PhpUnitaaa\InfrastructureTestCase;
 use Doctrine\ORM\EntityManager;
 
 abstract class MoocContextInfrastructureTestCase extends InfrastructureTestCase
