+++ tests/src/Shared/Infrastructure/Bus/Event/RabbitMq/RabbitMqEventBusTest.php
@@ -14,7 +14,7 @@ use CodelyTv\Shared\Infrastructure\Bus\Event\RabbitMq\RabbitMqEventBus;
 use CodelyTv\Shared\Infrastructure\Bus\Event\RabbitMq\RabbitMqQueueNameFormatter;
 use CodelyTv\Tests\Mooc\Courses\Domain\CourseCreatedDomainEventMother;
 use CodelyTv\Tests\Mooc\CoursesCounter\Domain\CoursesCounterIncrementedDomainEventMother;
-use CodelyTv\Tests\Shared\Infrastructure\PhpUnit\InfrastructureTestCase;
+use CodelyTv\Tests\Shared\Infrastructure\PhpUnitaaa\InfrastructureTestCase;
 use RuntimeException;
 use Throwable;
