+++ tests/src/Shared/Infrastructure/Bus/Query/InMemorySymfonyQueryBusTest.php
@@ -7,7 +7,7 @@ namespace CodelyTv\Tests\Shared\Infrastructure\Bus\Query;
 use CodelyTv\Shared\Domain\Bus\Query\Query;
 use CodelyTv\Shared\Infrastructure\Bus\Query\InMemorySymfonyQueryBus;
 use CodelyTv\Shared\Infrastructure\Bus\Query\QueryNotRegisteredError;
-use CodelyTv\Tests\Shared\Infrastructure\PhpUnit\UnitTestCase;
+use CodelyTv\Tests\Shared\Infrastructure\PhpUnitaaa\UnitTestCase;
 use RuntimeException;
 
 final class InMemorySymfonyQueryBusTest extends UnitTestCase
