+++ tests/src/Shared/Infrastructure/PHPUnitaaa/Comparator/AggregateRootArraySimilarComparator.php
@@ -2,9 +2,9 @@
 
 declare(strict_types = 1);
 
-namespace CodelyTv\Tests\Shared\Infrastructure\PhpUnit\Comparator;
+namespace CodelyTv\Tests\Shared\Infrastructure\PhpUnitaaa\Comparator;
 
-use CodelyTv\Shared\Domain\Bus\Event\DomainEvent;
+use CodelyTv\Shared\Domain\Aggregate\AggregateRoot;
 use CodelyTv\Tests\Shared\Domain\TestUtils;
 use SebastianBergmann\Comparator\Comparator;
 use SebastianBergmann\Comparator\ComparisonFailure;
@@ -12,14 +12,14 @@ use function Lambdish\Phunctional\all;
 use function Lambdish\Phunctional\any;
 use function Lambdish\Phunctional\instance_of;
 
-final class DomainEventArraySimilarComparator extends Comparator
+final class AggregateRootArraySimilarComparator extends Comparator
 {
     public function accepts($expected, $actual): bool
     {
         return is_array($expected) &&
                is_array($actual) &&
-               (all(instance_of(DomainEvent::class), $expected) &&
-                all(instance_of(DomainEvent::class), $actual));
+               (all(instance_of(AggregateRoot::class), $expected) &&
+                all(instance_of(AggregateRoot::class), $actual));
     }
 
     public function assertEquals($expected, $actual, $delta = 0.0, $canonicalize = false, $ignoreCase = false): void
@@ -31,16 +31,16 @@ final class DomainEventArraySimilarComparator extends Comparator
                 $this->exporter->export($expected),
                 $this->exporter->export($actual),
                 false,
-                'Failed asserting the collection of Events contains all the expected elements.'
+                'Failed asserting the collection of AGs contains all the expected elements.'
             );
         }
     }
 
     private function contains(array $expectedArray, array $actualArray): bool
     {
-        $exists = static function (DomainEvent $expected) use ($actualArray) {
+        $exists = function (AggregateRoot $expected) use ($actualArray) {
             return any(
-                static function (DomainEvent $actual) use ($expected) {
+                function (AggregateRoot $actual) use ($expected) {
                     return TestUtils::isSimilar($expected, $actual);
                 },
                 $actualArray
