+++ tests/src/Shared/Infrastructure/PHPUnitaaa/Comparator/DomainEventArraySimilarComparator.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace CodelyTv\Tests\Shared\Infrastructure\PhpUnit\Comparator;
+namespace CodelyTv\Tests\Shared\Infrastructure\PhpUnitaaa\Comparator;
 
 use CodelyTv\Shared\Domain\Bus\Event\DomainEvent;
 use CodelyTv\Tests\Shared\Domain\TestUtils;
