+++ tests/src/Shared/Infrastructure/PHPUnitaaa/Constraint/CodelyTvConstraintIsSimilar.php
@@ -2,14 +2,14 @@
 
 declare(strict_types = 1);
 
-namespace CodelyTv\Tests\Shared\Infrastructure\PhpUnit\Constraint;
+namespace CodelyTv\Tests\Shared\Infrastructure\PhpUnitaaa\Constraint;
 
-use CodelyTv\Tests\Shared\Infrastructure\PhpUnit\Comparator\AggregateRootArraySimilarComparator;
-use CodelyTv\Tests\Shared\Infrastructure\PhpUnit\Comparator\AggregateRootSimilarComparator;
-use CodelyTv\Tests\Shared\Infrastructure\PhpUnit\Comparator\DateTimeSimilarComparator;
-use CodelyTv\Tests\Shared\Infrastructure\PhpUnit\Comparator\DateTimeStringSimilarComparator;
-use CodelyTv\Tests\Shared\Infrastructure\PhpUnit\Comparator\DomainEventArraySimilarComparator;
-use CodelyTv\Tests\Shared\Infrastructure\PhpUnit\Comparator\DomainEventSimilarComparator;
+use CodelyTv\Tests\Shared\Infrastructure\PhpUnitaaa\Comparator\AggregateRootArraySimilarComparator;
+use CodelyTv\Tests\Shared\Infrastructure\PhpUnitaaa\Comparator\AggregateRootSimilarComparator;
+use CodelyTv\Tests\Shared\Infrastructure\PhpUnitaaa\Comparator\DateTimeSimilarComparator;
+use CodelyTv\Tests\Shared\Infrastructure\PhpUnitaaa\Comparator\DateTimeStringSimilarComparator;
+use CodelyTv\Tests\Shared\Infrastructure\PhpUnitaaa\Comparator\DomainEventArraySimilarComparator;
+use CodelyTv\Tests\Shared\Infrastructure\PhpUnitaaa\Comparator\DomainEventSimilarComparator;
 use PHPUnit\Framework\Constraint\Constraint;
 use PHPUnit\Framework\ExpectationFailedException;
 use SebastianBergmann\Comparator\ComparisonFailure;
